+++ examples/multithreaded.rs
@@ -81,15 +81,17 @@ fn main() {
                                     video_modes.iter().nth(video_mode_id).unwrap()
                                 );
                             }
-                            F => window.set_fullscreen(match (state, modifiers.alt()) {
-                                (true, false) => {
-                                    Some(Fullscreen::Borderless(window.current_monitor()))
-                                }
-                                (true, true) => Some(Fullscreen::Exclusive(
-                                    video_modes.iter().nth(video_mode_id).unwrap().clone(),
-                                )),
-                                (false, _) => None,
-                            }),
+                            F => window
+                                .set_fullscreen(match (state, modifiers.alt()) {
+                                    (true, false) => {
+                                        Some(Fullscreen::Borderless(window.current_monitor()))
+                                    }
+                                    (true, true) => Some(Fullscreen::Exclusive(
+                                        video_modes.iter().nth(video_mode_id).unwrap().clone(),
+                                    )),
+                                    (false, _) => None,
+                                })
+                                .unwrap(),
                             G => window.set_cursor_grab(state).unwrap(),
                             H => window.set_cursor_visible(!state),
                             I => {