+++ src/platform_impl/android/mod.rs
@@ -380,9 +380,10 @@ impl Window {
     }
 
     #[inline]
-    pub fn set_fullscreen(&self, _monitor: Option<RootMonitorHandle>) {
-        // N/A
-        // Android has single screen maximized apps so nothing to do
+    pub fn set_fullscreen(&self, _monitor: Option<RootMonitorHandle>) -> Result<(), Error> {
+        Err(make_error!(ErrorType::NotSupported(
+            "Android has single screen maximized apps so nothing to do".to_string()
+        )))
     }
 
     #[inline]